omit docker info ouput.
